Project did not compile

+++ src/main/java/net/journey/client/render/model/mob/boss/ModelSentryKing.java
@@ -136,9 +136,9 @@ public class ModelSentryKing extends ModelBase {
 	@Override
 	public void render(Entity entity, float f, float f1, float f2, float f3, float f4, float f5) {
 		super.render(entity, f, f1, f2, f3, f4, f5);
-		EntitySentryKing king = (EntitySentryKing)entity;
-		int healthDecimal = (int)king.getHealth() / king.getMaxHealth;
-		if(healthDecimal >= 0.9) {
+		EntitySentryKing king = (EntitySentryKing) entity;
+		int healthDecimal = (int) ((int) king.getHealth() / king.getMaxHealth());
+		if (healthDecimal >= 0.9) {
 			Head1.render(f5);
 			Head2.render(f5);
 			Head3.render(f5);
